+++ prac6/used_cars.py
@@ -2,9 +2,12 @@
 CP1404/CP5632 Practical - Client code to use the Car class.
 Note that the import has a folder (module) in it.
 This is why we name our folders with no spaces or capitals, as valid module names.
+
+estimated time: 20 mins
+actual time: 2 hours
 """
 
-from prac_06.car import Car
+from prac6.car import Car
 
 
 def main():
@@ -14,7 +17,7 @@ def main():
     print(f"Car has fuel: {my_car.fuel}")
     print(my_car)
 
-    my_limo = Car("limo",180)
+    my_limo = Car("limo", 180)
     my_limo.add_fuel(20)
     print(f"limo has fuel:{my_limo.fuel}")
     my_limo.drive(115)
